feat(RVCV-41): shopadmin의 리뷰 생성을 추가한다

+++ domain-module/review/src/main/java/com/romanticpipe/reviewcanvas/domain/Review.java
@@ -46,11 +46,11 @@ public class Review extends BaseEntityWithUpdate {
 	private ReviewStatus status;
 	private String imageVideoUrls;
 	private LocalDateTime deletedAt;
-	private Long shopAdminId;
+	private Integer shopAdminId;
 
 	@Builder
 	public Review(Long productId, User user, String content, int score, ReviewStatus status,
-		String imageVideoUrls, Long shopAdminId) {
+		String imageVideoUrls, Integer shopAdminId) {
 		this.productId = productId;
 		this.user = user;
 		this.replyList = new ArrayList<>();
